requirements.txt
python-telegram-bot==21.4
gspread==6.1.2
google-auth==2.33.0
openai==1.45.0
python-dotenv==1.0.1